+++ Core/Helper/Logfile.php
