--- xajax_core/Plugin/Loader.php
@@ -1,21 +1,0 @@
-<?php
-/**
- * PHP version php7
- *
- * @category
- * @package            xajax-php-7
- * @author             ${JProof}
- * @copyright          ${copyright}
- * @license            ${license}
- * @link
- * @see                ${docu}
- * @since              21.09.2017
- */
-
-declare(strict_types=1);
-
-namespace Xajax\Plugin;
-
-class Loader
-{
-}
