--- xajax_core/Plugin/PluginIface.php
@@ -1,21 +1,0 @@
-<?php
-/**
- * PHP version php7
- *
- * @category
- * @package            xajax-php-7
- * @author             ${JProof}
- * @copyright          ${copyright}
- * @license            ${license}
- * @link
- * @see                ${docu}
- * @since              15.10.2017
- */
-
-declare(strict_types=1);
-
-namespace Xajax\Plugin;
-
-interface PluginIface
-{
-}
